Lazy migration bug fixes
A  fix to solr_document behavior to find the correct model:

This enables us to have both models which are lazy migrating and
new models which are native Valkyrie in the same app.

A fix to the Freyja persister to handle lazy migration of members

The Freyja persister was submitting empty MigrateResourcesJobs because it
didn't check if there were any members to migrate before submitting.