Add commits from v5.0.4 back to main (#7023)
* Prevent loss of attributes with reschedule

This commit changes to use a temporary variable to hold the
job attributes, so that when CreateWorkJob is rescheduled due to
any errors, the attributes that were deleted do not get left out
of the resubmitted job's attributes.

* BatchCreateJob loses attributes

When the job is rescheduled, the attributes are lost.

This commit removes the reassignment of attributes so
a job autoretry has all of the attributes of the original job.

* Version bump to 5.0.4

* Fixing rubocop violation

* Fix chars eaten by merge conflict

---------

Co-authored-by: LaRita Robinson <larita@scientist.com>
Co-authored-by: LaRita Robinson <laritakr@users.noreply.github.com>
Co-authored-by: Daniel Pierce <dlpierce@iu.edu>