Add `Hyrax::FlexibleCatalogBehavior`
This commit will add the `Hyrax::FlexibleCatalogBehavior` which will
dynamically adjust the `CatalogController` according to the m3 profile.
If a property has 'stored_searchable' or 'something_tesim' in indexing,
then it will add it as a search term in all_fields.  If it has
'facetable' or 'something_sim', then it will add it as a facet on the
catalog results page.

Ref:
- https://github.com/notch8/wvu_knapsack/issues/35