[5.0 forward port] Valkyrie migration Improvements (#7082)
* Lazy migration bug fixes

A  fix to solr_document behavior to find the correct model:

This enables us to have both models which are lazy migrating and
new models which are native Valkyrie in the same app.

A fix to the Freyja persister to handle lazy migration of members

The Freyja persister was submitting empty MigrateResourcesJobs because it
didn't check if there were any members to migrate before submitting.

* Lazy migrate sipity entity with work

When lazy migration is used, the sipity entity cannot be found for works
which are not yet migrated. Valkyrie resources have a `proxy_for_global_id`
with the format:
`gid://hyku/Hyrax::ValkyrieGlobalIdProxy/2f6da5dd-9314-421f-b0dd-fda84fec9ee3`
while prior works used the model name such as:
`gid://hyku/GenericWork/2f6da5dd-9314-421f-b0dd-fda84fec9ee3`.

This commit introduces a lazy migration of the sipity entity via a job
which is submitted after a work is migrated to valkyrie via the Freyja
persister.

Prior work in pull request https://github.com/samvera/hyku/pull/2471
added some overrides in Hyku to resolve the sipity entity issues. These
changes still needs to be backported.

The prior work includes these changes:
The MigrateResourceService introduced a migration of the entity along
with the work's migration. This service is needed for direct migration
calls, but is not adequate for lazy migration, as the Freyja persister
does not call it for the work itself that was just migrated.

Hyku overrides to sipity.rb support cases where the entity is not yet migrated
and lazy migration is in use. This is necessary because the solr_document
is used to find the Sipity::Entity. Due to the model mapping found in the
solr document, the Entity method searches for the entity with the model
name. This is problematic when the entity is not yet migrated and the
entity is searched with a GenericWorkResource model rather than the
model.

* Convert id to string for job

* Don't rerun sipity job 5 times

We don't know until we try ActiveFedora whether there is might be a
problem with the sipity entry. We should not rerun the job 5 times
when the work was never in Fedora.

* couple of small things found while looking at save performance

* Catch errors that cause unnecessary retries

* Update app/jobs/migrate_files_to_valkyrie_job.rb

Co-authored-by: Rob Kaufman <rob@notch8.com>

* be more judicial when calling the migrate jobs

* use path that hyrax already creates

* make sure _path config options are paths and not just strings

* Rescue derivative migration failures (#7059)

* Handle nil file_set_params

In the controller, params[hash_key_for_curation_concern][:file_set]
resolves to nil when there are no files.

* fix blank string issue with file migration... blank files cant be stored by valkyrie

* call the cops

---------

Co-authored-by: LaRita Robinson <larita@scientist.com>
Co-authored-by: LaRita Robinson <larita@notch8.com>
Co-authored-by: LaRita Robinson <laritakr@users.noreply.github.com>
Co-authored-by: Randall Floyd <rdfloyd@iu.edu>