[5.0 forward port] Google analytics improvements (#7084)
* Fix parallel assignment bug and update partials

This commit will fix a parallel assignment bug where the syntax was not
doing what it thinks it was.  Also, the pagination partils were updated
with the blacklight theme to make it look more consistent with the rest
of the application.

* Add fixes for GA4

In the base, we are adding a few nil guards to we don't end up with a
NoMethodError for NilClass. In the GA4 module we are adding a case where
if we are getting a Hash already, don't try to JSON.parse it.  In the
UserStatImporter we are switching a query to the Valkyrie way, which
should already be in main.

* Handle invalid Google Analytics property ID gracefully

When the Google Analytics property ID is missing or invalid, catch the GRPC
and Google Cloud errors in the GA4 base service rather than letting them
bubble up to the view layer. This prevents dashboard errors and instead
returns empty analytics data when configuration is incomplete.

* reverts previous attempt to handle errors

default fall back to ga4

* refactor for maintainability and an easier read

* Rubocop fixes

* skip analytics_spec because google analytics provider is deprecated

---------

Co-authored-by: Kirk Wang <kirk.wang@scientist.com>
Co-authored-by: Shana Moore <shana.lavina.moore@gmail.com>
Co-authored-by: LaRita Robinson <larita@scientist.com>