Use solr by default in Hyrax::CustomQueries::FindIdsByModel (#7077)
* Adopt solrized find_ids_by_model from wings

* Match wings find_all_of_model query with count_all_of_model

Besides striving to be more consistent, this is needed in cases where a valkyrie model is converted to a wings model e.g. "Wings(Monograph)" but is still indexed with only its base model in has_model_ssim.

* Fix flaky spec

* Verify admin set more directly

The admin set model and the actual class representing it may not exactly match after running through wings. This changes the requirement from an exact match (after Wings model mapping) to any descendent of the registered admin set classes.

---------

Co-authored-by: Randall Floyd <rdfloyd@iu.edu>