Backports from main (#7117)
* Add depositor_ssim to core_metadata.yaml

FileSets were not getting the depositor_ssim indexed while other
resources were.  This then fails ::FileSet.search_in_batches search in
the UserStatImporter because `depositor_field` is set to
"depositor_ssim" in DepositSearchBuilder.

* Fix parallel assignment bug and update partials

This commit will fix a parallel assignment bug where the syntax was not
doing what it thinks it was.  Also, the pagination partils were updated
with the blacklight theme to make it look more consistent with the rest
of the application.

* Add fixes for GA4

In the base, we are adding a few nil guards to we don't end up with a
NoMethodError for NilClass. In the GA4 module we are adding a case where
if we are getting a Hash already, don't try to JSON.parse it.  In the
UserStatImporter we are switching a query to the Valkyrie way, which
should already be in main.

* Handle invalid Google Analytics property ID gracefully

When the Google Analytics property ID is missing or invalid, catch the GRPC
and Google Cloud errors in the GA4 base service rather than letting them
bubble up to the view layer. This prevents dashboard errors and instead
returns empty analytics data when configuration is incomplete.

* reverts previous attempt to handle errors

default fall back to ga4

* refactor for maintainability and an easier read

* Lazy migration bug fixes

A  fix to solr_document behavior to find the correct model:

This enables us to have both models which are lazy migrating and
new models which are native Valkyrie in the same app.

A fix to the Freyja persister to handle lazy migration of members

The Freyja persister was submitting empty MigrateResourcesJobs because it
didn't check if there were any members to migrate before submitting.

* Rubocop fixes

* Remove whitespace for rubocop

* 💄 removes trailing white space

* skip analytics_spec because google analytics provider is deprecated

* Lazy migrate sipity entity with work

When lazy migration is used, the sipity entity cannot be found for works
which are not yet migrated. Valkyrie resources have a `proxy_for_global_id`
with the format:
`gid://hyku/Hyrax::ValkyrieGlobalIdProxy/2f6da5dd-9314-421f-b0dd-fda84fec9ee3`
while prior works used the model name such as:
`gid://hyku/GenericWork/2f6da5dd-9314-421f-b0dd-fda84fec9ee3`.

This commit introduces a lazy migration of the sipity entity via a job
which is submitted after a work is migrated to valkyrie via the Freyja
persister.

Prior work in pull request https://github.com/samvera/hyku/pull/2471
added some overrides in Hyku to resolve the sipity entity issues. These
changes still needs to be backported.

The prior work includes these changes:
The MigrateResourceService introduced a migration of the entity along
with the work's migration. This service is needed for direct migration
calls, but is not adequate for lazy migration, as the Freyja persister
does not call it for the work itself that was just migrated.

Hyku overrides to sipity.rb support cases where the entity is not yet migrated
and lazy migration is in use. This is necessary because the solr_document
is used to find the Sipity::Entity. Due to the model mapping found in the
solr document, the Entity method searches for the entity with the model
name. This is problematic when the entity is not yet migrated and the
entity is searched with a GenericWorkResource model rather than the
model.

* Convert id to string for job

* Don't rerun sipity job 5 times

We don't know until we try ActiveFedora whether there is might be a
problem with the sipity entry. We should not rerun the job 5 times
when the work was never in Fedora.

* couple of small things found while looking at save performance

* Catch errors that cause unnecessary retries

* Update app/jobs/migrate_files_to_valkyrie_job.rb

Co-authored-by: Rob Kaufman <rob@notch8.com>

* be more judicial when calling the migrate jobs

* Queueable index adapter (#7053)

* working on a redis queue based index adapter.

* use path that hyrax already creates

* make sure _path config options are paths and not just strings

* Rescue derivative migration failures (#7059)

* Handle nil file_set_params

In the controller, params[hash_key_for_curation_concern][:file_set]
resolves to nil when there are no files.

* fix blank string issue with file migration... blank files cant be stored by valkyrie

* improve redis queue adapter initialization order and solve missing permisisons after indexing bug

* fix requeing

* redis indexing queue gets its own error queue

* error queue processing

* fix file metadata look up for valkyrie objects

---------

Co-authored-by: Kirk Wang <kirk.wang@scientist.com>
Co-authored-by: Kirk Wang <k3wang@gmail.com>
Co-authored-by: Shana Moore <shana.lavina.moore@gmail.com>
Co-authored-by: LaRita Robinson <larita@scientist.com>
Co-authored-by: LaRita Robinson <laritakr@users.noreply.github.com>
Co-authored-by: LaRita Robinson <larita@notch8.com>
Co-authored-by: Daniel Pierce <dlpierce@iu.edu>