Fix notification bell aria issues. (#7128)
* Fix notification bell aria issues.

aria-label did not match the displayed values. Moving that verbose string to aria-description and adding a label to the bell icon should resolve the issue and be more succinct.

The view spec is refactored to allow test messages to render, and some additional auth related specs are added.

* Attempt to quash trio of flaky ability specs

* Update notifications helper spec for new aria labels

* Try to fix more flaky sipity specs

* Move DatabaseCleaner.clean to later in the after hooks

An earlier hook waits for sessions to finish, so this should help capybara avoid hitting db cleaning deadlocks.

* Close any leftover db connections after cleaning.

May avoid flaky spec issues with sipity roles stragglers.