Pass query without using q:
Solr 9 breaks (unintentionally?) use of the terms query with the edismax query parser. The hyrax solrconfig.xml sets edismax as the default defType, but that is meant to be overridden back to lucene with `qt: standard` by `Hyrax::SolrService.additional_solr_get_and_post_params`. However, that override does not account for a query string passed via the `q:` named argument, as was the case here.