Fixes depositing a work into a collection
Depositing a work into a collection from the collection show page has been broken because the admin sets for selection were not being loaded. This caused the javascript to return "undefined" when trying to find the admin set, and caused the Hyrax.query_service.find to fail trying to find the admin set.

The fix encompasses several pieces:

Adding the available_admin_sets_for_creating_works method to the hyrax helper so it can be used in multiple places.
Updating the @admin_sets_for_select method in the Hyrax::My::WorksController to use the new helper method.
Adding the @admin_sets_for_select instance variable to the Hyrax::Dashboard::CollectionsController#show
Updating the javascript to return nil rather than undefined when no admin sets are available.