Fix required terms
The required terms were being pulled from the batch create form
rather than the actual selected work type. This has been fixed to ensure
that the correct required terms are used based on the selected work type.

Additionally, updates the CreateWorkJob to end with an error if the jobs
is not successful. Along with this, the job is modified to only run once.
Subsequent runs are not possible because uploaded_file is removed after
the first run, and any subsequent attempts will fail but override the
original cause for failure.