[Forward] 🐛 Collections facet in catalog search (#7190)
* 🐛 Collections facet in catalog search

This commit adds `member_of_collections_ssim` to the SolrDocument
ensuring collections with associated works are faceted in the catalog
search results.

Prior to this change, collections were faceted for Active Fedora works
but once the works were migrated or when new works are created the
`member_of_collections_ssim` was not present in the Solr document.

Ref:
- https://github.com/notch8/adventist_knapsack/issues/952

Co-Author: Sarah Proctor <sproctor950@gmail.com>

* spec fix