Allows migration of custom resource type
The issue we found was that the resource form couldn't be modified for a
custom worktype. This worked in the UI, but not in the migration job,
and the difference seemed to be that the form wasn't going through the same steps.

This fix mimics the steps taken through the UI when lazy migration happens.