Handle multiple files on a fileset
We've observed cases where the extracted text (which doesn't get created in Valkyrie)
does not have a filename and it overrides the original file.  The extracted text should
be in the derivatives, not with the file set.

With this change, we are just ignoring the files that don't have a filename. Current
version of Hyrax/Hyku do not use these secondary files on the fileset, so they are not
needed.