Fix collection creation with flexible metadata
This commit will address the bug where collection creation with flexible
metadata throws a `URI::InvalidURIError`.  This error is actually
misleading because what's happening is that when the form fails to
validate, it rerenders the new page but doesn't set the necessary
`collection_type_gid` that a normal new action would do.  The real
problem is that when using flexible metadata, validation fails because
depositor is not set in the form.

`Hyrax::FlexibleFormBehavior#validate_flexible_required_fields` (which
only runs when using flexible metadata) dynamically adds depositor as a
required field to the form instance.  Since depositor is not set in the
form params, validation fails.

This commit removes the property declaration on the `PcdmCollectionForm`
since this causes the error and matches `Hyrax::Forms::ResourceForm`
where it is not present and that's why creating a work does not trigger
the same error.