Remove Flexible Schema validation
For this branch, we are removing the validation that checks for
valid profile classes in the Flexible Schema model.

Having this validation in place was causing issues when creating
the default Flexible Schema while working in Hyrax development.
As the validations are handled in Hyku for this branch, they are not
needed here.

In the main branch, the validations are done differently and are not
affected by this change. This commit should NOT be backported to main.