🐛 Fix typo in valkyrie_transition? method call (#7295)
* 🐛 Fix typo in valkyrie_transition? method call

When not calling valkyrie_transition with a `?` it returns `nil` since
it's lazy loaded.  We want to call it with the `?` so it returns as
predicted.

* ✅ Bypass this spec

This commit will bypass this spec because it doesn't pass on Sirenia and
that is not relevent for our needs at all.  This is fixed in main but
5.0-flexible is not going into main so no need to fix it here.