Adds method `mappings_data_for` to FlexibleSchema
Adds new mappings_data_for method to extract mappings from profile to be used for OAI-PMH.

For OAI-PMH we need the mappings and indexing info for properties with the specified mapping. The result of this will be used to configure the OAI-PMH provider with the flexible metadata.

Includes a bugfix to correctly load hydra_model in the solr document when using valkyrie_transition.