🐛 Metadata properties respond to view key (#7279)
* 🐛 Metadata properties respond to view key

If a flexible metadata property has only display_label and admin_only
in the view hash, don't render the property on the catalog or show page.

Ref:
- https://github.com/samvera/hyku/issues/2774

* 🐛 Exclude required values for view options check

`display_label` and `admin_only` are always present in the view options
hash. This will ensure that only properties with additional specified
view options are passed on to the view.

This should be excluded:
```
{"display_label"=>{"default"=>"Depositor"}, "admin_only"=>false}
```

This should be included:
```
{"html_dl"=>true, "display_label"=>{"default"=>"Creator"}, "admin_only"=>false}
```

Ref:
- https://github.com/notch8/hyku-community-issues/issues/9

* ✅ Exclude properties without view options

If a schema (work type) has properties that do not have view options
other than the default `display_label` and `admin_only`, those
properties are excluded.

Ref:
- https://github.com/notch8/hyku-community-issues/issues/9

* ✅ Properties with view options render correctly

Ensures that properties with view options in addition to `admin_only`
and `display_label` are displayed as well as all deprecated label
options are removed.

Ref:
- https://github.com/notch8/hyku-community-issues/issues/9

* ✅ `view_definitions_for` return correct properties

The `view_definitions_for` returns the correct properties based on
contexts and view options.

Ref:
- https://github.com/notch8/hyku-community-issues/issues/9