Handle old and new style Deprecation#warn calls
The code (and dependencies) use a mix of signatures to call the
Deprecation#warn method. In some cases, this swallows the
text of the warning.

Patch the method to inspect arguments and ensure they get passed
in the correct positions to the older method provided by the gem.

```
DEPRECATION WARNING: You are using deprecated behavior which will be removed from the next major or minor release.
          (called from <class:Work> at /app/samvera/hyrax-engine/app/models/hyrax/work.rb:98)
```

```
DEPRECATION WARNING:  globally including `Hyrax::Schema(:core_metadata)` at the work level is deprecated.
          Add `include Hyrax::Schema(:core_metadata)` to your individual work classes or set
          up flexible metadata, then set `work_default_metadata = false` in config/hyrax.rb to remove this warning.
          (called from <class:Work> at /app/samvera/hyrax-engine/app/models/hyrax/work.rb:98)
```

* https://github.com/samvera/hyrax/issues/7303
* https://github.com/samvera/maintenance/issues/175