Make change set flexible
Moves flexible schema loading behavior into the ChangeSet class from the
ResourceForm class, so that it applies to all change sets, not just forms.

This fixes an issue where properties defined in a flexible schema were not
being set on the change set when not using a form from the workflow's
action taken service.